backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
sqlalchemy[asyncpg]>=2.0.0
asyncpg>=0.29.0
supabase>=2.3.1
google-generativeai>=0.3.0
psycopg2-binary==2.9.10
